Escape slashes in query_strings
Slashes (/) in query strings are treated by Elasticsearch as delimiting
a regexp expression. This is confusing especially when operating on
swift object paths, and in general we feel that in general the confusion
caused by mismatched slashes isn't offset by the ability to use regexps.
If a user really needs to run a regexp they can use the query dsl; we
could also consider adding a flag later.

Change-Id: I892541820b077d8eb08ae635a843d8dc84a5f55a
Closes-Bug: #1551946